feat: ensure cart product quantity can be decreased

+++ src/actions/decrease-cart-product-quantity/index.ts
@@ -8,30 +8,44 @@ import { db } from "@/db";
 import { cartItemTable } from "@/db/schema";
 import { auth } from "@/lib/auth";
 
-import { removeProductFromCartSchema } from "./schema";
+import { decreaseCartProductQuantitySchema } from "./schema";
 
-export const removeProductFromCart = async (
-  data: z.infer<typeof removeProductFromCartSchema>,
+export const decreaseCartProductQuantity = async (
+  data: z.infer<typeof decreaseCartProductQuantitySchema>,
 ) => {
-  removeProductFromCartSchema.parse(data);
+  decreaseCartProductQuantitySchema.parse(data);
   const session = await auth.api.getSession({
     headers: await headers(),
   });
+
   if (!session?.user) {
     throw new Error("Unauthorized");
   }
+
   const cartItem = await db.query.cartItemTable.findFirst({
     where: (cartItem, { eq }) => eq(cartItem.id, data.cartItemId),
     with: {
       cart: true,
     },
   });
+
   if (!cartItem) {
     throw new Error("Cart item not found");
   }
+
   const cartDoesNotBelongToUser = cartItem.cart.userId !== session.user.id;
+
   if (cartDoesNotBelongToUser) {
     throw new Error("Unauthorized");
   }
-  await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+
+  if (cartItem.quantity === 1) {
+    await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+    return;
+  }
+
+  await db
+    .update(cartItemTable)
+    .set({ quantity: cartItem.quantity - 1 })
+    .where(eq(cartItemTable.id, cartItem.id));
 };
